+++ apps/api/src/controllers/admin/manage/ManageUsersController.ts
@@ -22,7 +22,7 @@ import {
   PERMISSIONS_SCHEMA,
   ROLES_SCHEMA,
 } from "@snailycad/schemas";
-import { Socket } from "services/SocketService";
+import { Socket } from "services/socket-service";
 import { nanoid } from "nanoid";
 import { genSaltSync, hashSync } from "bcrypt";
 import { citizenInclude } from "controllers/citizen/CitizenController";
