+++ apps/api/src/controllers/admin/manage/manage-units-controller.ts
@@ -14,7 +14,7 @@ import { prisma } from "lib/prisma";
 import { validateSchema } from "lib/validateSchema";
 import { IsAuth } from "middlewares/IsAuth";
 import { UsePermissions, Permissions } from "middlewares/UsePermissions";
-import { Socket } from "services/SocketService";
+import { Socket } from "services/socket-service";
 import { ExtendedBadRequest } from "src/exceptions/ExtendedBadRequest";
 import { manyToManyHelper } from "utils/manyToMany";
 import { isCuid } from "cuid";
