+++ apps/api/src/controllers/calls/TowController.ts
@@ -12,7 +12,7 @@ import { prisma } from "lib/prisma";
 import { TOW_SCHEMA, UPDATE_TOW_SCHEMA } from "@snailycad/schemas";
 import { NotFound } from "@tsed/exceptions";
 import { IsAuth } from "middlewares/IsAuth";
-import { Socket } from "services/SocketService";
+import { Socket } from "services/socket-service";
 import { validateSchema } from "lib/validateSchema";
 import {
   cad,
