+++ apps/api/src/controllers/dispatch/911-calls/Calls911Controller.ts
@@ -9,7 +9,7 @@ import {
 import { HeaderParams, BodyParams, Context, PathParams, QueryParams } from "@tsed/platform-params";
 import { BadRequest, NotFound } from "@tsed/exceptions";
 import { prisma } from "lib/prisma";
-import { Socket } from "services/SocketService";
+import { Socket } from "services/socket-service";
 import { UseBeforeEach } from "@tsed/platform-middlewares";
 import { IsAuth } from "middlewares/IsAuth";
 import { _leoProperties } from "lib/leo/activeOfficer";
