+++ apps/api/src/lib/calls/assignUnitsToCall.ts
@@ -2,7 +2,7 @@ import { AssignedUnit, ShouldDoType } from "@prisma/client";
 import type { Call911 } from "@prisma/client";
 import { findUnit } from "lib/leo/findUnit";
 import { prisma } from "lib/prisma";
-import type { Socket } from "services/SocketService";
+import type { Socket } from "services/socket-service";
 import { manyToManyHelper } from "utils/manyToMany";
 import type { z } from "zod";
 import type { ASSIGNED_UNIT } from "@snailycad/schemas";
