+++ apps/api/src/services/socket-service.ts
